add a seperate slider component that comes with the text box
